Only admin can now remove and edit products.

+++ app/views/products/index.php
@@ -22,14 +22,16 @@
         <th class="align-middle text-center" scope="col">Fat <small>[g/100g]</small></th>
         <th class="align-middle text-center" scope="col">Carbs <small>[g/100g]</small></th>
         <th class="align-middle text-center" scope="col">Protein <small>[g/100g]</small></th>
-        <th class="align-middle" scope="col"></th>
-        <th class="align-middle" scope="col"></th>
+        <?php if(Session::currentUser()['role'] == 'admin') : ?>
+          <th class="align-middle" scope="col"></th>
+          <th class="align-middle" scope="col"></th>
+        <?php endif; ?>
       </tr>
     </thead>
     
     <tbody>
       <?php foreach($this->products as $product): ?>
-      <tr>
+        <tr>
           <td class="col-md-4">
             <a href="show/<?= $product->getId(); ?>"><?= $product->name; ?></a>
           </td>
@@ -39,13 +41,15 @@
           <td class="text-center"><?= $product->fat; ?></td>
           <td class="text-center"><?= $product->carbohydrates; ?></td>
           <td class="text-center"><?= $product->protein; ?></td>
-          <td class="col-md-1 text-center">
-            <a href="edit/<?= $product->getId(); ?>" class="text-body"><i class="bi bi-pencil-square"></i></a>
-          </td>
-          <td class="col-md-1 text-center">
-            <a href="delete/<?= $product->getId(); ?>" class="text-danger"><i class="bi bi-x-square"></i></a>
-          </td>
-      </tr>
+          <?php if(Session::currentUser()['role'] == 'admin') : ?>
+            <td class="col-md-1 text-center">
+              <a href="edit/<?= $product->getId(); ?>" class="text-body"><i class="bi bi-pencil-square"></i></a>
+            </td>
+            <td class="col-md-1 text-center">
+              <a href="delete/<?= $product->getId(); ?>" class="text-danger"><i class="bi bi-x-square"></i></a>
+            </td>
+          <?php endif; ?>
+        </tr>
       <?php endforeach; ?>
     </tbody>
   </table>
